.github/requirements.txt
moz-fluent-linter==0.2.*
compare-locales==8.2.1
